NAT-PMP private/public port and lifetime fix
